+++ routes/api/users.js
@@ -1,32 +1,7 @@
 const express = require("express");
-const { ctrlWrapper, auth } = require("../../middleWares");
+const { ctrlWrapper, auth, uploadFile } = require("../../middleWares");
 const { users: ctrl } = require("../../controllers");
-// const uploadFile = require("./uploadFile");
-// const upload = require("./UploadFile");
-
-const multer = require("multer");
-const path = require("path");
-const tempDir = path.join(process.cwd(), process.env.TEMP_DIR);
-
-const storage = multer.diskStorage({
-  destination: function (req, file, cb) {
-    cb(null, tempDir);
-  },
-  filename: function (req, file, cb) {
-    cb(null, file.originalname);
-  },
-});
-
-const upload = multer({
-  storage: storage,
-  limits: "10M",
-  fileFilter: (req, file, cb) => {
-    if (file.mimetype.includes("image")) {
-      cb(null, true);
-    }
-    cb(null, false);
-  },
-});
+// const uploadFile = require("../../middleWares/uploadFile");
 
 const router = express.Router();
 
@@ -41,16 +16,7 @@ router.get("/logout", ctrlWrapper(auth), ctrlWrapper(ctrl.logout));
 router.patch(
   "/avatars",
   ctrlWrapper(auth),
-  // uploadFile,
-  // ctrl.updateAvatar
-  // ctrlWrapper(uploadFile)
-  // UploadFile,
-  // (req, res, next) => uploadFile,
-  // uploadFile,
-  upload.single("avatar"),
-  // uploadFile.single("avatar"),
-  // uploadFile(req, res, next, "avatar")
-  // uploadFile.avatar,
+  uploadFile.single("avatar"),
   ctrlWrapper(ctrl.updateAvatar)
 );
 router.patch("/:userId", ctrlWrapper(auth), ctrlWrapper(ctrl.updateUser));
